Replace swith statement with a mapping
Official Python documentation explains [1], that switch statement can be
replaced by `if .. elif .. else` or by a mapping.

In first example with varying condtitions `if .. elif .. else` was used, and in
this example a mapping is much better fitted way of doing a switch statement
logic.

Since this is an official and documented way of doing switch statement logic in
Python, am adding it here.

[1] https://docs.python.org/3/faq/design.html#why-isn-t-there-a-switch-or-case-statement-in-python